Btrfs: add a helper to retrive extent inline ref type
An invalid value of extent inline ref type may be read from a
malicious image which may force btrfs to crash.

This adds a helper which does sanity check for the ref type, so we can
know if it's sane, return type if so, otherwise return an error.

v2: add enum type and return BTRFS_REF_TYPE_INVALID instead of -EINVAL

Signed-off-by: Liu Bo <bo.li.liu@oracle.com>