btrfs: add a qg_list abstraction for qgroup ulists
qgroups uses ulists as a heap allocated stack to workaround being unable
to use recursion to traverse the qgroup hierarchy. This is done from a
variety of common/important paths like space reservation, running
ordered_extents, running delayed_refs, and running qgroup stats.

This works great, except that it generally uses the qgroup spin lock and
GFP_ATOMIC. Except data reservation (but not metadata reservation), the
aforementioned callers do not have a syscall to report the error to, so
they result in aborted transactions and read only filesystems. This is a
problem that as previously fixed for extent_state allocations. All
callers allocate the ulist itself outside the spin lock with
GFP_KERNEL, but all of them use GFP_ATOMIC for the ulist_add calls.

To help with this issue, take advantage of the fact that in many
reasonable use cases, the qgroup hierarchy will not be particularly
deep. For example, by default it will always just consist of a single
subvol qgroup, but even with a few layers of nested qgroups, a qgroup
is unlikely to have tens of parent qgroups. With this assumption in
mind, add a qg_list abstraction that stack allocates an array of 8
btrfs_qgroup pointers and transparently uses either its array or a ulist
depending on the hierarchy size. If the hierarcy is <=8 qgroups, we will
never allocate for qgroup hierarchy operations.

It has a pretty simple api:
qg_list_init: init a new qg_list
qg_list_collect: collect the hierarchy for a qgroup into the qg_list
qg_list_iter_init: init an iterator on a qg_list
qg_list_iter_next: step through the qg_list_iter
qg_list_reinit: free per-collect allocations and reset for another collect
qg_list_free: free any allocated ulist memory

In my opinion, this also cleans up a lot of the ulist code to be less
repetitive and more readable. It does incur one cost, though, which is
that it no longer performs per-qgroup computations *while* collecting
them, but rather switches to collect, then compute.

i.e., in pseudocode:

old api, one loop:
ulist_add(ul, qg);
do_stuff(qg);
while (qg = get_next_ul_qg()) {
    do_stuff(qg);
    while (parent = get_next_qg_parent()) {
        ulist_add(ul, parent);
    }
}

becomes

new api, two loops:
qg_list_collect(qgl, qg); // loop over parents "recursively"
while (qg = qg_list_iter_next()) {
    do_stuff(qg);
}

We could theoretically pass in a computation to qg_list_collect to get
around this, but requiring it to be idempotent or the complexity of only
calling it once per qgroup in case we fall back to ulist makes me feel
it is not worth it.

Signed-off-by: Boris Burkov <boris@bur.io>