btrfs: convert qgroup_update_refcnt to qg_list
This function uses ulist in a relatively more complicated way. It has
two ulists, a per-qgroup collection ulist (tmp) and a longer-lived ulist
that collects all the qgroups across multiple calls to this function.
That second ulist is not modeled by the qg_list abstraction, so it
remains a plain ulist.

The new logic is roughly:
for root in roots:
    collect the hierarchy for that root's qg into tmp
    update refcnt for each collected qgroup
    save that qgroup in qgroups

we re-use the qg_list and tmp ulist throughout that for loop, so we use
qg_list_reinit explicity on each loop and qg_list_free on error or at
the end.

Signed-off-by: Boris Burkov <boris@bur.io>