btrfs: squota always reads eb in do_walk_down
There are cases where we can skip reading the next eb. However, for
simple quotas, we need the owner of the eb being derefed to accurately
account it. As-is, it is possible for us to not have next read in, skip
reading it in, and try to read from it (btrfs_header_owner(next) when
preparing the generic_ref). This can fail in a variety of ways, but
usually by trying to release metadata space from an invalid qgroup which
doesn't exist.

If simple quotas are enabled, never skip reading next. Perhaps in the
future we can come up with conditions we can still skip in, but this is
the safest for now.

This reproduced for me on a vm with 2G of ram running btrfs/187.

Signed-off-by: Boris Burkov <boris@bur.io>