btrfs: zoned: lock page for when dirtying extent buffer
Josef reported a crash in fstests' generic/163 on zoned devices,
caused by an ASSERT()ion. This ASSERT() is checking if a page is set
dirty.

Like in the subpage case, lock the extent buffer's page when setting
it dirty.

This prevents racing with other threads cleaning the dirty bit and
hitting the ASSERT() later on.

Reported-by: Josef Bacik <josef@toxicpanda.com>
Signed-off-by: Johannes Thumshirn <johannes.thumshirn@wdc.com>