btrfs: forbid creating subvol qgroups
This leads to various races and it isn't helpful, because you can't
specify a subvol id when creating a subvol, so you can't be sure it will
be the right one. Any requirements on the automatic subvol can be
gratified by using a higher level qgroup and the inheritance parameters
of subvol creation.

CC: stable@vger.kernel.org # 4.14+
Reviewed-by: Qu Wenruo <wqu@suse.com>
Signed-off-by: Boris Burkov <boris@bur.io>
Reviewed-by: David Sterba <dsterba@suse.com>
Signed-off-by: David Sterba <dsterba@suse.com>