btrfs: zoned: fix conversion of zoned filesystems to RAID
When running a balance convert on a zoned filesystem, it is currently
possible to select an unsupported profile. 

The first patch in this series fixes this issue and is indented to be
backported to kernel version that do not have the RAID stripe tree
feature.

The second patch fixes the issue for RAID stripe tree enabled filesystems.

The third patch is independent of this series but, something I came
across when working on this series.

To: Chris Mason <clm@fb.com>
To: Josef Bacik <josef@toxicpanda.com>
To: David Sterba <dsterba@suse.com>
To: Naohiro Aota <naohiro.aota@wdc.com>
To: Anand Jain <anand.jain@oracle.com>
Cc:  <linux-btrfs@vger.kernel.org>
Cc:  <linux-kernel@vger.kernel.org>


--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 1,
    "change-id": "20240207-rst-convert-fix-15deefc4753a",
    "prefixes": []
  }
}