btrfs: zoned: traverse device list under RCU lock in btrfs_zoned_should_reclaim()
As btrfs_zoned_should_reclaim() traverses the device list with the
device_list_mutex held. But we're never changing the device list. All we
do is gathering the used and total bytes.

So change the list traversal from the holding the device_list_mutex to
rcu_read_lock(). This also opens up the possibilities to call
btrfs_zoned_should_reclaim() with the chunk_mutex held.

Signed-off-by: Johannes Thumshirn <johannes.thumshirn@wdc.com>