btrfs: zoned: always set aside a zone for relocation
For zoned filesytsems we heavily rely on relocation for garbage collecting
as we cannot do any in-place updates of disk blocks.

But there can be situations where we're running out of space for doing the
relocation.

To solve this, always have a zone reserved for relocation.

This is a subset of another approach to this problem I've submitted in 
https://lore.kernel.org/r/20240328-hans-v1-0-4cd558959407@kernel.org

To: Chris Mason <clm@fb.com>
To: Josef Bacik <josef@toxicpanda.com>
To: David Sterba <dsterba@suse.com>
CC: Hans Holmberg <Hans.Holmberg@wdc.com>
Cc:  <linux-btrfs@vger.kernel.org>
Cc:  <linux-kernel@vger.kernel.org>

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 1,
    "change-id": "20240514-zoned-gc-2ce793459eb7",
    "prefixes": []
  }
}