btrfs: rst: updates for RAID stripe tree
Patch 1 replaces stripe extents in case we hit a EEXIST when inserting a
stripe extent on a write. This can happen i.e. on device-replace.

Patch 2 splits a stripe extent on partial delete of a stripe.

Patch 3 adds selftests for the stripe-tree delete code, these selftests
can and will be extended to cover insert and get as well.

Patch 4 fixes a deadlock between scrub and device replace when RST is
used.

Patch 5 get's rid of the pointless tree dump in case we're not finding an
RST entry.

Patch 6 is a prep-patch form #7 and renames btrfs_io_stripe::is_scrub to
commit_root.

Patch 7 changes the read code so we're looking at the commit_root for
relocation as well.

To: Chris Mason <clm@fb.com>
To: Josef Bacik <josef@toxicpanda.com>
To: David Sterba <dsterba@suse.com>
Cc:  <linux-btrfs@vger.kernel.org>
Cc:  <linux-kernel@vger.kernel.org>

---
Changes in v5:
- EDITME: describe what is new in this series revision.
- EDITME: use bulletpoints and terse descriptions.
- Link to v4: https://lore.kernel.org/r/20240705-b4-rst-updates-v4-0-f3eed3f2cfad@kernel.org

Changes in v4:
- Addressed Josef's comments
- Added patches 6 & 7
- Link to v3: https://lore.kernel.org/r/20240701-b4-rst-updates-v3-0-e0437e1e04a6@kernel.org

Changes in v3:
- Drop on-disk format change as it's in for-next
- Add patches 4 & 5
- Link to v2: https://lore.kernel.org/r/20240619-b4-rst-updates-v2-0-89c34d0d5298@kernel.org

Changes in v2:
- Added selftests for delete code
- Link to v1: https://lore.kernel.org/r/20240610-b4-rst-updates-v1-0-179c1eec08f2@kernel.org

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 5,
    "change-id": "20240610-b4-rst-updates-d0aa696b9d5a",
    "prefixes": [],
    "history": {
      "v1": [
        "20240610-b4-rst-updates-v1-0-179c1eec08f2@kernel.org"
      ],
      "v2": [
        "20240619-b4-rst-updates-v2-0-89c34d0d5298@kernel.org"
      ],
      "v3": [
        "20240701-b4-rst-updates-v3-0-e0437e1e04a6@kernel.org"
      ],
      "v4": [
        "20240705-b4-rst-updates-v4-0-f3eed3f2cfad@kernel.org"
      ]
    }
  }
}