btrfs: rst: updates for RAID stripe tree
Three independent updates for RAID stripe tree.
The 1st one removes pointless space from the on-disk format. As the
feature itself is still experimental I'd like to get rid of that as early
as possible.

Patch 2 replaces stripe extents in case we hit a EEXIST when inserting a
stripe extent on a write. This can happen i.e. on device-replace.

Patch 3 splits a stripe extent on partial delete of a stripe.

Patch 4 adds selftests for the stripe-tree delete code, these selftests
can and will be extended to cover insert and get as well

To: Chris Mason <clm@fb.com>
To: Josef Bacik <josef@toxicpanda.com>
To: David Sterba <dsterba@suse.com>
Cc:  <linux-btrfs@vger.kernel.org>
Cc:  <linux-kernel@vger.kernel.org>

---
Changes in v3:
- EDITME: describe what is new in this series revision.
- EDITME: use bulletpoints and terse descriptions.
- Link to v2: https://lore.kernel.org/r/20240619-b4-rst-updates-v2-0-89c34d0d5298@kernel.org

Changes in v2:
- Added selftests for delete code
- Link to v1: https://lore.kernel.org/r/20240610-b4-rst-updates-v1-0-179c1eec08f2@kernel.org

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 3,
    "change-id": "20240610-b4-rst-updates-d0aa696b9d5a",
    "prefixes": [],
    "history": {
      "v1": [
        "20240610-b4-rst-updates-v1-0-179c1eec08f2@kernel.org"
      ],
      "v2": [
        "20240619-b4-rst-updates-v2-0-89c34d0d5298@kernel.org"
      ]
    }
  }
}