btrfs: tree-checker: skip name hash check for image dump
[BUG]
For an image dumpped with "btrfs-image -s", the restored filesystem will
not pass tree-checker:

 BTRFS critical (device dm-5): corrupt leaf: root=1 block=32522240 slot=6 ino=6, name hash mismatch with key, have 0x0000000019e196de expect 0x000000008dbfc2d2
 BTRFS error (device dm-5): read time tree block corruption detected on logical 32522240 mirror 1

[CAUSE]
Btrfs-image with "-s" option would sanitize the filenames, thus it's
ensured to cause file names to mismatch their hash.

[FIX]
Since btrfs-image is mostly for debug purposes, we can afford it to be
an exception for tree-checker.

Just allow image dump to skip the name hash mismatch.

There will be a little more risk, as image dump can still be mounted RW,
but the existing error handling is good enough to handle the mismatched
hash.

Reported-by: Andrea Gelmini <andrea.gelmini@gmail.com>
Signed-off-by: Qu Wenruo <wqu@suse.com>