btrfs: do not use __GFP_NOFAIL flag for btrfs_alloc_folio_array()
Since all callers of this function is already properly handling the
allocation error, and for the worst case we're just going to abort the
current transaction, I do no believe we need __GFP_NOFAIL here.

So to enable more testing and hopefully to provide a smooth transaction,
for CONFIG_BTRFS_DEBUG builds do not use __GFP_NOFAIL flag.

Signed-off-by: Qu Wenruo <wqu@suse.com>