btrfs: more RAID stripe tree updates
Three further RST updates targeted for 6.11 (hopefully).

The first one is a reworked version of the scrub vs dev-replace deadlock
fix. It does have reviews from Josef and Qu but I'd love to head Filipe's
take on it.

The second one updates a stripe extent in case a write to a already
present logical address happens.

The third breaks out of the delete loop once we're past the range we want
to delete.

To: Chris Mason <clm@fb.com>
To: Josef Bacik <josef@toxicpanda.com>
To: David Sterba <dsterba@suse.com>
Cc:  <linux-btrfs@vger.kernel.org>
Cc:  <linux-kernel@vger.kernel.org>
Cc: Qu Wenru <wqu@suse.com>
Cc: Filipe Manana <fdmanana@suse.com>

---
Changes in v2:
- Add Qu's Reviewed-by on patch 2
- Add patch 3
- Link to v1: https://lore.kernel.org/r/20240709-b4-rst-updates-v1-0-200800dfe0fd@kernel.org

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 2,
    "change-id": "20240709-b4-rst-updates-bb9c0e49cd5b",
    "prefixes": [],
    "history": {
      "v1": [
        "20240709-b4-rst-updates-v1-0-200800dfe0fd@kernel.org"
      ]
    }
  }
}